# openapi.yaml
openapi: 3.0.3
info:
  title: CyberShield API
  description: API для образовательного симулятора кибербезопасности. Платформа обучает пользователей основам кибербезопасности через интерактивные сценарии, миссии и систему геймификации.
  version: 1.0.0
  contact:
    name: CyberShield Team
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.cybershield.example.com
    description: Production server
...
paths:
...
  /api/scenarios/:
    get:
      summary: Список всех сценариев
      tags:
        - Scenarios
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список сценариев с прогрессом
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScenarioListItem'
  /api/scenarios/{slug}:
    get:
      summary: Детали сценария с миссиями
      tags:
        - Scenarios
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: Уникальный slug сценария
      responses:
        '200':
          description: Детали сценария
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioResponse'
        '404':
          description: Сценарий не найден
  /api/missions/{mission_id}:
    get:
      summary: Получение миссии с шагами и вариантами ответов
      tags:
        - Missions
      security:
        - BearerAuth: []
      parameters:
        - name: mission_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID миссии
      responses:
        '200':
          description: Детали миссии
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionDetail'
        '404':
          description: Миссия не найдена
  /api/missions/{mission_id}/start:
    post:
      summary: Начало миссии
      tags:
        - Missions
      security:
        - BearerAuth: []
      parameters:
        - name: mission_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Миссия начата
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionStartResponse'
        '404':
          description: Миссия не найдена
  /api/missions/{mission_id}/answer:
    post:
      summary: Отправка ответа на шаг миссии
      tags:
        - Missions
      security:
        - BearerAuth: []
      parameters:
        - name: mission_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerRequest'
      responses:
        '200':
          description: Результат ответа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerResponse'
        '400':
          description: Ошибка валидации
  /api/missions/{mission_id}/complete:
    post:
      summary: Завершение миссии
      tags:
        - Missions
      security:
        - BearerAuth: []
      parameters:
        - name: mission_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Результаты завершения миссии
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionResultResponse'
        '400':
          description: Ошибка
  /api/progress/:
    get:
      summary: Сводка прогресса пользователя
      tags:
        - Progress
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Сводка прогресса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressSummary'
  /api/progress/dashboard:
    get:
      summary: Данные для дашборда
      tags:
        - Progress
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Полный дашборд
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardData'
  /api/progress/history:
    get:
      summary: История прохождения миссий
      tags:
        - Progress
      security:
        - BearerAuth: []
      responses:
        '200':
          description: История миссий
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistoryItem'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token. Получите через POST /api/auth/login или POST /api/auth/register.
  schemas:
...
    ScenarioListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
        order_index:
          type: integer
        is_active:
          type: boolean
        missions_count:
          type: integer
        completed_count:
          type: integer
    ScenarioResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        difficulty:
          type: integer
        order_index:
          type: integer
        is_active:
          type: boolean
        missions:
          type: array
          items:
            $ref: '#/components/schemas/MissionListItem'
    MissionListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        attack_type:
          type: string
          description: Тип атаки (phishing, skimming, password, social)
        environment:
          type: string
          description: Среда (email, browser, mobile, etc.)
        difficulty:
          type: integer
        xp_reward:
          type: integer
        order_index:
          type: integer
        time_limit_seconds:
          type: integer
          nullable: true
        status:
          type: string
          enum: [not_started, in_progress, completed]
    MissionDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        story_text:
          type: string
          nullable: true
        attack_type:
          type: string
        environment:
          type: string
        difficulty:
          type: integer
        xp_reward:
          type: integer
        time_limit_seconds:
          type: integer
          nullable: true
        steps:
          type: array
          items:
            $ref: '#/components/schemas/MissionStepResponse'
    MissionStepResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        step_order:
          type: integer
        context_text:
          type: string
          description: Текст контекста/ситуации
        context_type:
          type: string
          description: Тип контекста (story, email, message, etc.)
        context_data:
          type: object
          nullable: true
          description: Дополнительные данные контекста
        choices:
          type: array
          items:
            $ref: '#/components/schemas/StepChoiceResponse'
    StepChoiceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        choice_text:
          type: string
        order_index:
          type: integer
    MissionStartResponse:
      type: object
      properties:
        mission:
          type: object
          properties:
            id:
              type: string
              format: uuid
            slug:
              type: string
            title:
              type: string
            story_text:
              type: string
              nullable: true
            attack_type:
              type: string
            environment:
              type: string
            xp_reward:
              type: integer
            time_limit_seconds:
              type: integer
              nullable: true
        first_step:
          $ref: '#/components/schemas/MissionStepResponse'
          nullable: true
        current_hp:
          type: integer
          default: 100
        current_step:
          type: integer
          default: 0
    AnswerRequest:
      type: object
      required:
        - step_id
        - choice_id
      properties:
        step_id:
          type: string
          format: uuid
        choice_id:
          type: string
          format: uuid
    AnswerResponse:
      type: object
      properties:
        is_correct:
          type: boolean
        feedback_text:
          type: string
          description: Обратная связь по ответу
        consequence_text:
          type: string
          nullable: true
          description: Описание последствий выбора
        hp_change:
          type: integer
          description: Изменение HP
        current_hp:
          type: integer
        next_step:
          $ref: '#/components/schemas/MissionStepResponse'
          nullable: true
        mission_completed:
          type: boolean
        new_achievements:
          type: array
          items:
            $ref: '#/components/schemas/AchievementNotification'
    MissionResultResponse:
      type: object
      properties:
        xp_earned:
          type: integer
        correct_answers:
          type: integer
        wrong_answers:
          type: integer
        total_steps:
          type: integer
        accuracy_percent:
          type: number
          format: float
        new_achievements:
          type: array
          items:
            $ref: '#/components/schemas/AchievementNotification'
        league_changed:
          type: boolean
        new_league:
          type: string
          nullable: true
    ProgressSummary:
      type: object
      properties:
        total_missions:
          type: integer
        completed_missions:
          type: integer
        total_xp:
          type: integer
        security_level:
          type: integer
        accuracy_percent:
          type: number
          format: float
        total_errors:
          type: integer
    DashboardData:
      type: object
      properties:
        user_stats:
          $ref: '#/components/schemas/ProgressSummary'
        recent_achievements:
          type: array
          items:
            $ref: '#/components/schemas/AchievementNotification'
        scenario_progress:
          type: array
          items:
            type: object
            properties:
              slug:
                type: string
              title:
                type: string
              icon:
                type: string
              total:
                type: integer
              completed:
                type: integer
        attack_type_stats:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              total:
                type: integer
              completed:
                type: integer
    HistoryItem:
      type: object
      properties:
        mission_id:
          type: string
          format: uuid
        mission_title:
          type: string
        scenario_title:
          type: string
        status:
          type: string
        xp_earned:
          type: integer
        correct_answers:
          type: integer
        wrong_answers:
          type: integer
        completed_at:
          type: string
          format: date-time
          nullable: true
...
    AchievementNotification:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        rarity:
          type: string
        xp_bonus:
          type: integer
        earned_at:
          type: string
          format: date-time
...